Reorganized into a header/source structure
