+++ src/PoseCorrectionNode.hpp
@@ -1,0 +1,161 @@
+#include "pose_estimation.hpp"
+#include <apriltag_msgs/msg/april_tag_detection.hpp>
+#include <apriltag_msgs/msg/april_tag_detection_array.hpp>
+#ifdef cv_bridge_HPP
+#include <cv_bridge/cv_bridge.hpp>
+#else
+#include <cv_bridge/cv_bridge.h>
+#endif
+#include <image_transport/camera_subscriber.hpp>
+#include <image_transport/image_transport.hpp>
+#include <rclcpp/rclcpp.hpp>
+#include <rclcpp_components/register_node_macro.hpp>
+#include <sensor_msgs/msg/camera_info.hpp>
+#include <sensor_msgs/msg/image.hpp>
+#include <tf2_ros/static_transform_broadcaster.h>
+#include <tf2_ros/transform_broadcaster.h>
+#include <tf2_ros/buffer.h>
+#include <tf2_ros/transform_broadcaster.h>
+#include <tf2_ros/transform_listener.h>
+#include <tf2_geometry_msgs/tf2_geometry_msgs.hpp>
+// #include <zed_msgs/srv/set_pose.hpp>
+#include <zed_interfaces/srv/set_pose.hpp>
+#include <Eigen/Dense>
+
+
+// LINK - Class Definition
+//Class to encompass the apriltag detection and pose correction
+class PoseCorrectionNode : public rclcpp::Node
+{
+    public:
+   
+        PoseCorrectionNode(const rclcpp::NodeOptions& options);
+
+        ~PoseCorrectionNode();
+
+    private:
+
+        // Callback handle to allow updating parameters during runtime
+        const OnSetParametersCallbackHandle::SharedPtr paramCallbackHandler_;
+
+        // Apriltag family and detector pointers
+        std::string tagFamilyStr_;
+        apriltag_family_t* tagFamily_;
+        apriltag_detector_t* const tagDetector_;
+
+        // parameter
+        std::mutex mutex_;
+        double tagEdgeSize_;
+        std::atomic<int> maxHamming_;
+        std::atomic<bool> profile_;
+        std::unordered_map<int, std::string> tagFrames_;
+        std::unordered_map<int, double> tagSizes_;
+        std::vector<long int> tagIDs_;
+
+        // Maps of tag IDs to pose and frame name
+        std::unordered_map<int64_t, std::vector<_Float64>> globalTagPoseMap_;
+        std::unordered_map<int64_t, std::string> tagIDtoFrame_;
+
+        // TF destructor
+        std::function<void(apriltag_family_t*)> tagFamilyDestructor_;
+
+        // Camera subscription, AprilTag Detections, and broadcasters for transformations
+        const image_transport::CameraSubscriber cameraSub_;
+        const rclcpp::Publisher<apriltag_msgs::msg::AprilTagDetectionArray>::SharedPtr detectionPub_;
+        tf2_ros::TransformBroadcaster tfBroadcaster_;
+        tf2_ros::StaticTransformBroadcaster staticBroadcaster_;
+
+        rclcpp::Publisher<geometry_msgs::msg::PoseWithCovarianceStamped>::SharedPtr posePub_;
+
+        // Transform Buffer and Listener
+        std::shared_ptr<tf2_ros::Buffer> transformBuffer_{nullptr};
+        std::shared_ptr<tf2_ros::TransformListener> transformListener_;
+        
+
+        // Service client to reset pose
+        rclcpp::Client<zed_interfaces::srv::SetPose>::SharedPtr setPoseClient_;
+
+        // Function used to estimate pose
+        pose_estimation_f estimatePose_ = nullptr;
+
+        std::string cameraName_;
+        std::string worldFrame_;
+
+        // Transforms to move between reference frames
+        tf2::Transform camLeftToBase_;
+        tf2::Transform imageToROS_;
+        tf2::Transform rosToImage_;
+        tf2::Transform tagToImage_;
+        tf2::Transform imageToTag_;
+        tf2::Transform camCenterToBaseLink_;
+
+
+        // tf2::Transform apriltagToCam_;
+        // tf2::Transform camToApriltag_;
+        // tf2::Transform camToRos_;
+        // tf2::Transform rosToCam_;
+
+        // Map of tag frames to transforms
+        std::unordered_map<std::string, tf2::Transform> tagTransformMap_;
+
+
+
+
+        /**
+         * @brief Callback to detect Apriltags and call the pose correction
+         * Service
+         * 
+         * @param img Image from the ZED X camera
+         * @param camInfo Camera info from the ZED X camera
+         */
+        void onCamera(
+            const sensor_msgs::msg::Image::ConstSharedPtr & img,
+            const sensor_msgs::msg::CameraInfo::ConstSharedPtr & camInfo);
+
+        /**
+         * @brief Initializes transformations
+         * 
+         */
+        void initTFs();
+
+        /**
+         * @brief Get existing transformations from TF. This function is
+         * pulled from the ZED ArUco tag localization example. 
+         * 
+         * @param targetFrame 
+         * @param sourceFrame 
+         * @param out_tr 
+         */
+        bool getTransformFromTf(
+            std::string targetFrame, std::string sourceFrame,
+            tf2::Transform & out_tr);
+        
+        /**
+         * @brief Calls the set_pose service from the ZED ROS2 Wrapper
+         * to correct the pose.
+         * 
+         * @param new_pose 
+         * @return true 
+         * @return false 
+         */
+        bool resetZedPose(tf2::Transform & new_pose);
+
+        /**
+         * @brief Takes the transform msg provided by estiamte_pose
+         * 
+         * @param tf - Transform msg for the closest tag
+         * @param id - ID of the closest tag
+         */
+        void computeTransform(tf2::Transform & tf, int id);
+    
+        /**
+         * @brief Callback that is triggered when parameters are changed. Pulled from AprilTagNode
+         * 
+         * @param parameters 
+         * @return rcl_interfaces::msg::SetParametersResult 
+         */
+        rcl_interfaces::msg::SetParametersResult onParameter(const std::vector<rclcpp::Parameter>& parameters);
+
+
+};
+RCLCPP_COMPONENTS_REGISTER_NODE(PoseCorrectionNode)
